% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{b4t_levels}
\alias{b4t_levels}
\title{Factor levels for the B4T scenario column}
\format{
A character vector.
}
\usage{
b4t_levels
}
\description{
The six CG regions plus "All CG regions" (the CG6 scenario).
}
\keyword{datasets}
