% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{b4t_regions}
\alias{b4t_regions}
\title{CGIAR regions where Breeding For Tomorrow (B4T) is implemented}
\format{
A character vector of CG region codes.
}
\usage{
b4t_regions
}
\description{
CGIAR regions where Breeding For Tomorrow (B4T) is implemented
}
\keyword{datasets}
