% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b4t.R
\name{build_b4t_comparisons}
\alias{build_b4t_comparisons}
\title{B4T comparisons for a list of indicator specs}
\usage{
build_b4t_comparisons(specs, ...)
}
\arguments{
\item{specs}{Indicator specs (see \code{\link[=indicator_specs]{indicator_specs()}}).}

\item{...}{Further arguments to \code{\link[=build_b4t_comparison]{build_b4t_comparison()}}.}
}
\value{
All comparisons stacked, with an \code{indicator} column.
}
\description{
B4T comparisons for a list of indicator specs
}
