% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMolecules.R
\name{countMolecules}
\alias{countMolecules}
\title{Make a count matrix from molecule-level information}
\usage{
countMolecules(
  num.features,
  feature,
  barcode,
  count = NULL,
  class = c("dgCMatrix", "SVT_SparseMatrix"),
  num.threads = 1
)
}
\arguments{
\item{num.features}{Integer specifying the total number of features used in the reference annotation.}

\item{feature}{Integer vector of length equal to the number of molecules, containing the feature to which molecule was assigned.
A value greater than \code{num.features} indicates that the corresponding molecule is unassigned.}

\item{barcode}{Factor of length equal to the number of molecules, containing the cell barcode to which each molecule was assigned.
Levels are usualy barcode sequences.}

\item{class}{String specifying the output class.}

\item{num.threads}{Integer specifying the number of threads to use.}

\item{counts}{Integer vector of length equal to the number of molecules, containing the read counts for each molecule.
If supplied, the count matrix refers to the number of reads assigned to each gene in each barcode, not the number of molecules (i.e., UMIs).}
}
\value{
A sparse matrix of the specified \code{class}.
Each row is a feature and each column is a barcode.
Each entry is the number of molecules (if \code{counts=NULL}) or reads (otherwise) for the corresponding feature and barcode.
Columns are named after the corresponding levels of \code{barcode}.
}
\description{
Construct a count matrix from per-molecule information, typically the barcode and feature to which each molecule was assigned.
}
\examples{
# Mocking up some 10X molecule information.
sim <- simulateMoleculeInformation(num.features=20)

# Formatting it into a matrix.
countMolecules(20, sim$molecules$feature, sim$molecules$barcode)

}
\seealso{
\code{\link{readMoleculeInformation}}, which is typically used to load molecule information from the HDF5 file produced by CellRanger.
}
\author{
Aaron Lun
}
