% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{crio_baseline_id}
\alias{crio_baseline_id}
\title{Default baseline scenario}
\format{
A character string.
}
\usage{
crio_baseline_id
}
\description{
SSP3, MRI climate model, RCP 7.0, no CO2 fertilization.
}
\keyword{datasets}
