% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSequences.R
\name{encodeSequences}
\alias{encodeSequences}
\alias{decodeSequences}
\title{Encode and decode sequences}
\usage{
encodeSequences(sequences)

decodeSequences(encoded, lengths)
}
\arguments{
\item{sequences}{Character vector of nucleotide sequences.}

\item{encoded}{Integer vector of 2-bit encodings of the sequences.}

\item{lengths}{Integer vector containing the lengths of sequences.
This should be the same length as \code{encoded}, otherwise it is recycled to the length of \code{encoded}.}
}
\value{
For \code{encodeSequences}, an integer vector containing the 2-bit integer encoding for each sequence.

For \code{decodeSequences}, a character vector containing the decoded sequences.
}
\description{
Encode sequences to a 2-bit integer representation, and decode the integers back to sequences.
}
\examples{
enc <- encodeSequences(c("AAAA", "CCCC", "TTTT", "GGGG"))
enc
decodeSequences(enc, lengths=4)

}
\seealso{
\code{\link{readMolecules}}, which uses a 2-bit encoding for the UMI sequences.
}
\author{
Aaron Lun
}
