% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{food_groups}
\alias{food_groups}
\title{Food groups used for calorie and water totals}
\format{
A character vector of GFSir food group codes.
}
\usage{
food_groups
}
\description{
Food groups used for calorie and water totals
}
\keyword{datasets}
