% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_foodavail}
\alias{get_foodavail}
\title{Food availability (all commodities and food groups)}
\usage{
get_foodavail(files, mapping = "mappingCG6.xlsx")
}
\arguments{
\item{files}{Scenario GDX files (see \code{\link[=list_scenarios]{list_scenarios()}}).}

\item{mapping}{GFSir mapping file name.}
}
\value{
GFSir \code{foodavail} output for all scenarios (no scenario columns).
}
\description{
Food availability (all commodities and food groups)
}
