% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_frtveg}
\alias{get_frtveg}
\title{Fruit and vegetable availability (g/person/day)}
\usage{
get_frtveg(foodavail)
}
\arguments{
\item{foodavail}{Output of \code{\link[=get_foodavail]{get_foodavail()}}.}
}
\value{
F&V availability.
}
\description{
Fruit and vegetable availability (g/person/day)
}
