% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_mder}
\alias{get_mder}
\title{Total calories against the minimum dietary energy requirement (MDER)}
\usage{
get_mder(calories, minkcal)
}
\arguments{
\item{calories}{Output of \code{\link[=get_calories]{get_calories()}}.}

\item{minkcal}{Output of \code{\link[=get_minkcal]{get_minkcal()}}.}
}
\value{
Total calories with \code{minKCAL} and \code{hits_target}.
}
\description{
Total calories against the minimum dietary energy requirement (MDER)
}
