% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_sugar_share}
\alias{get_sugar_share}
\title{Share of calories from sugar}
\usage{
get_sugar_share(calories, sugar_name = "O&S-Sugar  (p)")
}
\arguments{
\item{calories}{Output of \code{\link[=get_calories]{get_calories()}}.}

\item{sugar_name}{Commodity name for refined sugar.}
}
\value{
One row per region x year x scenario with \code{sugar_energy_share}.
}
\description{
Share of calories from sugar
}
