% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{load_region_map}
\alias{load_region_map}
\title{Load the IMPACT region shapefile}
\usage{
load_region_map(shp)
}
\arguments{
\item{shp}{Path to the IMPACT regions shapefile (with a \code{NEW_REGION} column).}
}
\value{
A SpatVector.
}
\description{
Load the IMPACT region shapefile
}
