% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\docType{data}
\name{nutrient_labels}
\alias{nutrient_labels}
\title{Display names for nutrient codes}
\format{
A named character vector, names are \code{Nutrients} codes.
}
\usage{
nutrient_labels
}
\description{
Display names for nutrient codes
}
\keyword{datasets}
