% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{plot_pct_hits}
\alias{plot_pct_hits}
\title{Bar chart of the share of scenarios hitting a target}
\usage{
plot_pct_hits(analysis, title = "\% scenarios hitting target")
}
\arguments{
\item{analysis}{Output of \code{\link[=build_pathway_analysis]{build_pathway_analysis()}}.}

\item{title}{Plot title.}
}
\value{
A ggplot.
}
\description{
Regions where some but not all scenarios hit the target.
}
