% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathways.R
\name{plot_region_tree}
\alias{plot_region_tree}
\title{Plot the classification tree for one region}
\usage{
plot_region_tree(analysis, region_code, title = NULL)
}
\arguments{
\item{analysis}{Output of \code{\link[=build_pathway_analysis]{build_pathway_analysis()}}.}

\item{region_code}{Region code.}

\item{title}{Plot title.}
}
\value{
Called for its side effect (base graphics plot).
}
\description{
Plot the classification tree for one region
}
