% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCounts.R
\name{readCounts}
\alias{readCounts}
\alias{configureSampleForReadCounts}
\title{Load data from a 10X Genomics experiment}
\usage{
readCounts(
  samples,
  column.names = FALSE,
  row.names = c("id", "symbol"),
  delayed = FALSE,
  intersect.rows = FALSE,
  BPPARAM = NULL,
  ...
)

configureSampleForReadCounts(
  path,
  type = NULL,
  version = NULL,
  genome = NULL,
  compressed = NULL,
  mtx.two.pass = FALSE,
  mtx.class = c("CsparseMatrix", "SVT_SparseMatrix"),
  mtx.threads = 1,
  hdf5.realize = FALSE,
  hdf5.realize.class = c("CsparseMatrix", "SVT_SparseMatrix")
)
}
\arguments{
\item{samples}{A character vector of paths.
Each entry will be interpreted as described below for \code{path} with an automatically-detected \code{type}.

Alternatively, a list of sample information.
Each element corresponds to a sample and is either a list returned by \code{configureSampleForReadCounts} or a string containing a path.

In both cases, the list or vector can be named, in which case the names are used to identify each sample in the output object.
If unnamed, the sample identity is set to the path instead.

Users may also pass the output of a single call to \code{configureSampleForReadCounts}. 
This is treated as a list of length 1 containing the configuration of the lone sample.}

\item{column.names}{Boolean indicating whether the columns of the output object should be named with the cell barcodes.}

\item{row.names}{String specifying whether to use Ensembl ids ("id") or gene symbols ("name") as row names. 
For symbols, the Ensembl id will be appended to disambiguate rows where the same symbol corresponds to multiple Ensembl ids.}

\item{delayed}{Boolean indicating whether sparse matrices should be wrapped in \link[DelayedArray]{DelayedArray}s before combining.
This saves memory by avoiding the creation of a new matrix; it also avoids integer overflow for dgCMatrix instances.
Only relevant for multiple \code{samples}.}

\item{intersect.rows}{Boolean indicating whether to take the intersection of common row names across all samples.
If \code{FALSE}, differences in row names across samples will cause an error to be raised.}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object specifying how loading should be parallelized for multiple \code{samples}.
If \code{NULL}, no parallelization is performed.}

\item{...}{Further arguments to pass to \code{configureSampleForReadCounts} when any entry of \code{samples} is a string representing a path.
All such paths in \code{samples} are internally configured with these parameters.}

\item{path}{String specifying the path to the CellRanger output for a single sample.
\itemize{
\item If \code{type="hdf5"}, \code{path} should be a HDF5 file in the 10X sparse HDF5 format. 
Users may need to set \code{genome} if it cannot be automatically determined when \code{version="2"}.
\item If \code{type="mtx"}, \code{path} should refer to a directory containing:
\itemize{
\item \code{"matrix.mtx"}, a Matrix Market file containing a feature-by-barcode UMI count matrix.
\item \code{"barcodes.tsv"}, a tab-delimited file containing per-barcode information.
\item \code{"features.tsv"}, a tab-delimited file containing information for each genomic feature (usually genes).
For \code{version="2"}, this will be called \code{"genes.tsv"} instead.
}
Any of these files may be Gzip-compressed, in which case they should have an additional \code{".gz"} file extension.
\item If \code{type="prefix"}, \code{path} should be a path prefix for all files associated with this sample.
The files should follow the same format described above for \code{type="mtx"}.
For example, if \code{path="xyz_"}, the files associated with this sample should be \code{"xyz_matrix.mtx"}, \code{"xyz_barcodes.tsv"}, etc.
Again, any of these files may be Gzip-compressed, in which case they should be suffixed with \code{".gz"}.
}}

\item{type}{String specifying the type of 10X format to read data from.
If \code{NULL}, the function will attempt to automatically detect this from the specified files -
\code{"hdf5"} if the path ends with \code{".h5"}, \code{"mtx"} if the path refers to a directory, and \code{"prefix"} otherwise.}

\item{version}{String specifying the version of the 10X format to read data from.
If \code{NULL}, the function will attempt to automatically detect this from the specified files.
For \code{type="mtx"}, this is based on whether the per-feature information is stored in a \code{"features.tsv*"} or \code{"genes.tsv*"} file.
For \code{type="hdf5"}, this is based on whether there is a top-level \code{"matrix"} group with a \code{"matrix/features"} subgroup in the file.}

\item{genome}{String specifying the genome, i.e., the HDF5 group containing the data.
If \code{NULL}, the function will attempt to automatically detect this from the specified files.
Only used if \code{type="hdf5"} and \code{version='2'}.}

\item{compressed}{Boolean indicating whether the text files are compressed for \code{type="mtx"} or \code{"prefix"}.
If \code{NULL}, the function will attempt to automatically detect this from the specified files, i.e., whether the file paths end with \code{".gz"}.}

\item{mtx.two.pass}{Boolean indicating whether to use a two-pass approach for loading data from a Matrix Market file.
This reduces peak memory usage at the cost of some additional runtime. 
Only relevant when \code{type="mtx"} or \code{type="prefix"}.}

\item{mtx.class}{String specifying the class of the output matrix when \code{type="mtx"} or \code{type="prefix"}.}

\item{mtx.threads}{Boolean specifying the number of threads to use for reading Matrix Market files.
Only relevant when \code{type="mtx"} or \code{type="prefix"}.}

\item{hdf5.realize}{Boolean specifying whether the count data should be loaded into memory for \code{type="hdf5"}.
If \code{FALSE}, the count matrix is represented as a file-backed matrix.}

\item{hdf5.realize.class}{String specifying the class of the output matrix when \code{type="hdf5"} and \code{hdf5.realize=TRUE}.}
}
\value{
For \code{readCounts}, a \link[SingleCellExperiment]{SingleCellExperiment} object containing count data for each genomic feature (row) and cell (column) across all \code{samples}.
\itemize{
\item The \code{\link[SummarizedExperiment]{rowData}} contains the \code{"id"} and \code{"name"} columns.
The former is the gene identifier (usually Ensembl), while the latter is the gene name.
If \code{version="3"}, it will also contain the \code{"type"} field specifying the type of feature (e.g., gene, antibody tag, CRISPR guide).
\item The \code{\link[SummarizedExperiment]{rowRanges}} contains the genomic coordinates for each gene.
Only provided for \code{version="3"}, 
\item The \code{\link[SummarizedExperiment]{colData}} will contain the \code{"sample"} and \code{"barcode"} columns.
The former contains the name of the sample (or if not supplied, the path in \code{samples}) from which each column was obtained.
The latter contains the cell barcode sequence and GEM group for each cell library. 
\item Rows are named with the gene identifier.
If \code{column.names=TRUE}, each column is named by the cell barcode if there is only a single sample.
For multiple samples, the index of each sample in \code{samples} is concatenated to the cell barcode to form the column name.
This avoids problems with multiple instances of the same cell barcodes in different samples.
\item The \code{\link[SummarizedExperiment]{assays}} will contain a single \code{"counts"} matrix, containing UMI counts for each gene in each cell in each sample.
Matrices are combined by column if multiple \code{samples} were specified.
Note that the matrix representation will depend on the format of the \code{samples}:
\itemize{
\item For \code{type="prefix"} or \code{type="mtx"}, the class of the matrix is determined by \code{mtx.class}.
For \code{type="hdf5"}, the matrix is a \link[HDF5Array]{HDF5Matrix} if \code{hdf5.realize=FALSE}, otherwise its class is determined by \code{hdf5.realize.class}.
\item If multiple samples are present, the class of the matrix is determined by the return type of \code{cbind} on the per-sample matrices.
This can be forced to be a \link[DelayedArray]{DelayedMatrix} by setting \code{delayed=TRUE}. 
}
\item The \code{\link[S4Vectors]{metadata}} contains a \code{"samples"} field, containing the input \code{samples} character vector.
}

For \code{configureSampleForReadCounts}, a list of configuration parameters for each sample.
This should be passed to \code{samples} in \code{readCounts}.
}
\description{
Creates a \link[SingleCellExperiment]{SingleCellExperiment} from the CellRanger output directories for 10X Genomics data.
}
\details{
This function has a long and storied past.
It was originally developed as the \code{read10xResults} function in \pkg{scater}, inspired by the \code{Read10X} function from the \pkg{Seurat} package.
It was then migrated to \pkg{DropletUtils} in an effort to consolidate some 10X-related functionality across various packages.
Finally, it was moved to \pkg{crio} to separate the input/output utilities from analysis functions.

Note that user-level manipulation of sparse matrices requires loading of the \pkg{Matrix} package.
Otherwise, calculation of \code{rowSums}, \code{colSums}, etc. will result in errors.
}
\examples{
# Mocking up some 10X genomics output.
example(writeCounts, echo=FALSE)

# Reading it in.
sce10x <- readCounts(tmpdir)

# Works with multiple samples.
sce10x2 <- readCounts(c(tmpdir, tmpdir))

}
\references{
Zheng GX, Terry JM, Belgrader P, and others (2017).
Massively parallel digital transcriptional profiling of single cells. 
\emph{Nat Commun} 8:14049.

10X Genomics (2017).
Gene-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/output/matrices}

10X Genomics (2018).
Feature-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices}

10X Genomics (2018).
HDF5 Gene-Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/advanced/h5_matrices}

10X Genomics (2018).
HDF5 Feature Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices}
}
\seealso{
\code{\link[SingleCellExperiment]{splitAltExps}}, to split alternative feature sets (e.g., antibody tags) into their own Experiments.

\code{\link{writeCounts}}, to create 10X-formatted file(s) from a count matrix.
}
\author{
Davis McCarthy, with modifications from Aaron Lun.
}
