% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMoleculeInformation.R
\name{readMoleculeInformation}
\alias{readMoleculeInformation}
\title{Read the 10X molecule information file}
\usage{
readMoleculeInformation(
  path,
  barcode.length = NULL,
  keep.unmapped = FALSE,
  get.barcode = TRUE,
  get.umi = TRUE,
  get.gem = TRUE,
  get.feature = TRUE,
  get.count = TRUE,
  get.library = TRUE,
  extract.library.info = TRUE,
  version = NULL
)
}
\arguments{
\item{path}{String containing the path to the molecule information HDF5 file.}

\item{barcode.length}{Integer specifying the length of the cell barcode.
If \code{NULL}, the barcode length is guessed from the 2-bit encoded integers. 
Only relevant when \code{version="2"}.}

\item{keep.unmapped}{Boolean indicating whether unmapped molecules should be reported.}

\item{get.barcode, get.umi, get.gem, get.feature, get.count, get.library}{Logical scalar indicating whether the corresponding field should be extracted for each molecule.}

\item{extract.library.info}{Boolean whether the library information should be extracted.
Only relevant when \code{version="3"}.}

\item{version}{String specifying the version of the 10X molecule information format to read data from.
This can be either \code{"2"} or \code{"3"}.
If \code{NULL}, the version is auto-detected from the contents of \code{path}.}
}
\value{
A named list containing molecule information.

The first entry of this list is \code{molecules}, a \link[S4Vectors]{DataFrame} where each row corresponds to a single transcript molecule.
This contains the following fields:
\describe{
\item{\code{Barcode}:}{Factor, the cell barcode for each molecule.
The levels contain the universe of all known barcodes.}
\item{\code{Umi}:}{Integer, the processed UMI barcode in 2-bit encoding.} 
\item{\code{GemGroup}:}{Integer, the GEM group.}
\item{\code{Feature}:}{Integer, the index of the feature to which the molecule was assigned.
This refers to an entry in the \code{features} DataFrame, see below.}
\item{\code{Count}:}{Integer, the number of reads mapped to this molecule.}
\item{\code{Library}:}{Integer, the library index in cases where multiple libraries are present in the same file.
Only reported when \code{version="3"}.}
}
A field will not be present in the DataFrame if the corresponding \code{get.*} argument is \code{FALSE}, 

The second element of the list is \code{features}, a \link[S4Vectors]{DataFrame} where each row corresponds to a genomic feature.
This contains the following fields:
\describe{
\item{\code{ID}:}{Character, the identifier for the feature.
For genes, this is typically the Ensembl ID.}
\item{\code{Type}:}{Character, the type of the feature, e.g., \code{"Gene Expression"}, \code{"Antibody Capture"}.}
}
This is indexed by the \code{Feature} field in the \code{data} DataFrame.

If \code{extract.library.info=TRUE}, the list will contain an additional element named \code{libraries}. 
This is a list of lists containing per-library information such as the \code{"library_type"}.
The \code{library} field in the \code{data} DataFrame indexes this list.
}
\description{
Extract relevant fields from the molecule information HDF5 file.
}
\details{
Molecules that were not assigned to any gene have their \code{Feature} set to \code{nrow(features) + 1}.
These are removed when \code{keep.unmapped = FALSE}.

CellRanger 3.0 introduced a major change in the format of the molecule information files.
When \code{version="auto"}, the function will attempt to determine the version format of the file.
This can also be user-specified by setting \code{version} explicitly.

For files produced by version 2.2 of the CellRanger software, the length of the cell barcode is not given.
Instead, the barcode length is automatically inferred if \code{barcode.length=NULL} and \code{version="2"}.
Currently, version 1 of the 10X chemistry uses 14 nt barcodes, while version 2 uses 16 nt barcodes.

Setting any of the \code{get.*} arguments will (generally) avoid extraction of the corresponding field.
This can improve efficiency if that field is not necessary for further analysis.
Aside from the missing field, the results are guaranteed to be identical, i.e., same order and number of rows.
}
\examples{
# Mocking up some 10X HDF5-formatted data.
sim <- simulateMoleculeInformation()
temp <- tempfile(fileext=".h5")
writeMoleculeInformation(temp, sim)

# Reading the resulting file.
readMoleculeInformation(temp)

}
\references{
Zheng GX, Terry JM, Belgrader P, and others (2017).
Massively parallel digital transcriptional profiling of single cells. 
\emph{Nat Commun} 8:14049.

10X Genomics (2017).
Molecule info.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/output/molecule_info}

10X Genomics (2018).
Molecule info.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/molecule_info}
}
\seealso{
\code{\link{makeCountMatrix}}, which creates a count matrix from this information.
}
\author{
Aaron Lun,
based on code by Jonathan Griffiths
}
