% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_data.R
\docType{data}
\name{rni_lookup}
\alias{rni_lookup}
\title{Recommended nutrient intakes (RNI)}
\format{
A tibble with columns \code{Nutrients} and \code{RNI}.
}
\usage{
rni_lookup
}
\description{
Adult RNIs per capita per day, from Sherwin's GTAP paper. Units are in the
nutrient code suffix (g, mg, mcg).
}
\keyword{datasets}
