% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{row_countries}
\alias{row_countries}
\title{Rest-of-world countries greyed out on maps}
\format{
A character vector of IMPACT region codes.
}
\usage{
row_countries
}
\description{
Rest-of-world countries greyed out on maps
}
\keyword{datasets}
