% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_id}
\alias{scenario_id}
\title{Scenario id from a GDX file path}
\usage{
scenario_id(gdx)
}
\arguments{
\item{gdx}{GDX file path(s).}
}
\value{
File name without folder and \code{.gdx} extension.
}
\description{
Scenario id from a GDX file path
}
