% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCounts.R
\name{writeCounts}
\alias{writeCounts}
\title{Write count data in the 10x format}
\usage{
writeCounts(
  path,
  x,
  assay = 1L,
  force.integer = TRUE,
  barcode.field = NULL,
  feature.id.field = NULL,
  feature.name.field = "name",
  feature.type.field = "type",
  overwrite = FALSE,
  type = NULL,
  version = "3",
  genome = "unknown",
  compressed = (version == "3"),
  chemistry = "Single Cell 3' v3",
  original.gem.groups = 1L,
  library.ids = "custom",
  num.threads = 1
)
}
\arguments{
\item{path}{A string containing the path to the output directory (for \code{type="mtx"}) or file (for \code{type="HDF5"}).}

\item{x}{A \link[SummarizedExperiment]{SummarizedExperiment} containing a matrix of UMI counts.}

\item{assay}{Integer or string specifying the assay of \code{x} containing the count matrix to be written.}

\item{barcode.field}{String containing the name of the \code{\link[SummarizedExperiment]{colData}} column containing the cell barcodes.
If \code{NULL}, the column names of \code{x} are assumed to contain the barcodes.}

\item{feature.id.field}{String containing the name of the \code{\link[SummarizedExperiment]{rowData}} column containing the feature IDs.
If \code{NULL}, the row names of \code{x} are assumed to contain the feature IDs.}

\item{feature.name.field}{String containing the name of the \code{\link[SummarizedExperiment]{rowData}} column containing the feature names.
If \code{NULL}, the row names of \code{x} are assumed to contain the feature names.}

\item{feature.type.field}{String containing the name of the \code{\link[SummarizedExperiment]{rowData}} column containing the feature types.
If this name is not present in the \code{rowData}, the type defaults to \code{"Gene Expression"} for all rows of \code{x}.
Only used when \code{version="3"}.}

\item{overwrite}{Boolean specifying whether \code{path} should be overwritten if it already exists.}

\item{type}{String specifying the type of 10X format to save \code{x} to.
This is either a directory containing a sparse matrix with row/column annotation (\code{"mtx"})
or a HDF5 file containing the same information (\code{"HDF5"}).}

\item{version}{String specifying the version of the CellRanger format to produce.}

\item{genome}{String specifying the genome for storage when \code{type="HDF5"}.
This can be a character vector with one genome per feature if \code{version="3"}.}

\item{chemistry, original.gem.groups, library.ids}{Strings containing metadata attributes to be added to the HDF5 file for \code{type="HDF5"}.
Their interpretation is not formally documented and is left to the user's imagination.}

\item{num.threads}{Integer specifying the number of threads to use for counting the number of non-zero elements.}

\item{force.integers}{Boolean specifying whether entries of the count matrix should be truncated to their integer component.
This can be set to \code{FALSE} to preserve non-integer values in the on-disk representation.}
}
\value{
For \code{type="mtx"}, a directory is produced at \code{path}.
If \code{version="2"}, this will contain the files \code{"matrix.mtx"}, \code{"barcodes.tsv"} and \code{"genes.tsv"}.
If \code{version="3"}, it will instead contain \code{"matrix.mtx.gz"}, \code{"barcodes.tsv.gz"} and \code{"features.tsv.gz"}.

For \code{type="HDF5"}, a HDF5 file is produced at \code{path} containing data in column-sparse format.
If \code{version="2"}, data are stored in the HDF5 group named \code{genome}.
If \code{version="3"}, data are stored in the group \code{"matrix"}.

A \code{TRUE} value is invisibly returned.
}
\description{
Create a directory containing the count matrix and cell/gene annotation from a sparse matrix of UMI counts, 
in the format produced by the CellRanger software suite.
}
\details{
This function will try to automatically detect the desired format based on whether \code{path} ends with \code{".h5"}.
If so, it assumes that \code{path} specifies a HDF5 file path and sets \code{type="HDF5"}.
Otherwise it will set \code{type="mtx"} under the assumption that \code{path} specifies a path to a directory.

Note that there were major changes in the output format for CellRanger version 3.0 to account for non-gene features such as antibody or CRISPR tags. 
Users can switch to this new format using \code{version="3"}.
See the documentation for \dQuote{latest} for this new format, otherwise see \dQuote{2.2} or earlier.

The primary purpose of this function is to create files to use for testing \code{\link{read10xCounts}}.
In principle, it is possible to re-use the HDF5 matrices in \code{cellranger reanalyze}.
We recommend against doing so routinely due to CellRanger's dependence on undocumented metadata attributes that may change without notice.
}
\examples{
# Mocking up some count data.
library(Matrix)
my.counts <- matrix(rpois(1000, lambda=5), ncol=10, nrow=100)
my.counts <- as(my.counts, "dgCMatrix")

library(SummarizedExperiment)
se <- SummarizedExperiment(my.counts)
colnames(se) <- paste0("BARCODE-", seq_len(ncol(my.counts)))
rownames(se) <- paste0("ENSG0000", seq_len(nrow(my.counts)))
rowData(se)$name <- paste0("GENE", seq_len(nrow(my.counts)))
rowData(se)$type <- "Gene Expression"

# Writing this to file:
tmpdir <- tempfile()
writeCounts(tmpdir, se)
list.files(tmpdir)

# Creating a version 3 HDF5 file:
tmph5 <- tempfile(fileext=".h5")
writeCounts(tmph5, se, version='3')

}
\references{
10X Genomics (2017).
Gene-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/output/matrices}

10X Genomics (2018).
Feature-Barcode Matrices.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices}

10X Genomics (2018).
HDF5 Gene-Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/2.2/advanced/h5_matrices}

10X Genomics (2018).
HDF5 Feature Barcode Matrix Format.
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/advanced/h5_matrices}
}
\seealso{
\code{\link{read10xCounts}}, to read CellRanger matrices into R.
}
\author{
Aaron Lun
}
