% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMolecules.R
\name{writeMolecules}
\alias{writeMolecules}
\alias{simulateMolecules}
\title{Write the 10X molecule information file}
\usage{
writeMolecules(path, data, version = c("3", "2"))

simulateMolecules(
  num.molecules = 10000,
  barcode.length = 4,
  num.features = 20,
  num.gem.groups = 1,
  umi.length = 10,
  ave.num.reads = 10,
  version = c("3", "2")
)
}
\arguments{
\item{path}{String containing the path to the file.}

\item{data}{List returned by \code{\link{readMolecules}} or \code{\link{simulateMolecules}}.
This should contain the \code{"molecules"}, \code{"features"} and (for \code{version="3"}) \code{"libraries"} entries.}

\item{version}{String specifying the version of the 10X molecule information format to write.
This can be either \code{"2"} or \code{"3"}.}

\item{num.molecules}{Integer specifying the number of molecules to simulate.}

\item{barcode.length}{Integer specifying the length of the cell barcodes.}

\item{num.features}{Integer specifying the number of features in the simulation.}

\item{num.gem.groups}{Integer specifying the number of GEM groups. 
This is also treated as the number of libraries for \code{version="3"}.}

\item{umi.length}{Integer specifying the length of the UMI.}

\item{ave.num.reads}{Number specifying the average of the number of reads per molecule.}
}
\value{
For \code{writeMolecules}, a new HDF5 file is created at \code{path} containing the molecule information.
\code{NULL} is invisibly returned.

For \code{simulateMolecules}, a list is returned containing \code{molecules}, \code{features}, and (for \code{version="3"}) \code{libraries}.
Each of these mimic the corresponding entry in the list returned by \code{\link{readMolecules}}.
}
\description{
Write a data frame of molecule information into a HDF5 file.
}
\details{
These functions are primarily intended for examples and testing of \code{readMolecules}.
The files written by \code{writeMolecules} may not be sufficiently compliant with 10X's specification for use in external applications.
If you are running into interoperability issues, file an issue at \url{https://github.com/LTLA/crio} and we'll see what we can do.
}
\examples{
sim <- simulateMolecules()
tmp <- tempfile(fileext=".h5")
writeMolecules(tmp, sim)
rhdf5::h5ls(tmp)

}
\author{
Aaron Lun
}
