% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crio-package.R
\docType{package}
\name{crio-package}
\alias{crio}
\alias{crio-package}
\title{crio: CellRanger Input/Output}
\description{
Load CellRanger's output files into Bioconductor data structures, most typically a SingleCellExperiment containing assay values. Also supports writing to CellRanger formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LTLA/crio}
  \item Report bugs at \url{https://github.com/LTLA/crio/issues}
}

}
\author{
\strong{Maintainer}: Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
}

Other contributors:
\itemize{
  \item Jonathan Griffiths [contributor]
}

}
